requirements/app.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/app.txt requirements/app.in
#
aniso8601==1.3.0          # via flask-restful
blinker==1.4              # via raven
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via flask
contextlib2==0.5.5        # via raven
flask-marshmallow==0.8.0
flask-restful==0.3.6
flask==0.12.2
gunicorn==19.7.1
idna==2.6                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
markupsafe==1.0           # via jinja2
marshmallow==2.13.6       # via flask-marshmallow
python-dateutil==2.6.1    # via aniso8601
pytz==2017.2              # via flask-restful
pyyaml==3.12
raven[flask]==6.2.1
requests==2.18.4
six==1.11.0               # via flask-marshmallow, flask-restful, python-dateutil
urllib3==1.22             # via requests
werkzeug==0.12.2          # via flask
